requirements.txt
mosaicml[nlp,wandb]==0.32.1
mosaicml-streaming==0.7.6
transformers==4.40.2
python-dotenv==1.0.0
huggingface-hub>=0.20.0
tqdm==4.67.1